Move CLI encoding to application package
